changement categorieproduit

+++ src/main/java/tn/esprit/rh/achat/entities/CategorieProduit.java
@@ -26,7 +26,7 @@ public class CategorieProduit implements Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer idCategorieProduit;
+	private Long idCategorieProduit;
 	private String codeCategorie;
 	private String libelleCategorie;
 	@OneToMany(mappedBy = "categorieProduit")
